[next] Update app test fixture
