+++ packages/next/test/fixtures/00-app-dir/app/(newroot)/dashboard/another/page.js
