+++ packages/next/test/fixtures/00-app-dir/app/(rootonly)/dashboard/hello/page.js
