+++ packages/next/test/fixtures/00-app-dir/app/client-component-route/page.js
@@ -1,3 +1,5 @@
+'client';
+
 import { useState, useEffect } from 'react';
 
 import style from './style.module.css';
