+++ packages/next/test/fixtures/00-app-dir/app/client-nested/layout.js
@@ -1,3 +1,5 @@
+'client';
+
 import { useState, useEffect } from 'react';
 
 import styles from './style.module.css';
