+++ packages/next/test/fixtures/00-app-dir/app/dashboard/deployments/[id]/page.js
@@ -1,0 +1,7 @@
+export default function DeploymentsPage(props) {
+  return (
+    <>
+      <p>hello from app/dashboard/deployments/[id]. ID is: {props.params.id}</p>
+    </>
+  );
+}
