+++ packages/next/test/fixtures/00-app-dir/app/dashboard/deployments/layout.js
@@ -1,0 +1,8 @@
+export default function DeploymentsLayout({ message, children }) {
+  return (
+    <>
+      <h2>Deployments hello</h2>
+      {children}
+    </>
+  );
+}
