--- packages/next/test/fixtures/00-app-dir/app/dashboard/deployments/layout.server.js
@@ -1,16 +1,0 @@
-export function getServerSideProps() {
-  return {
-    props: {
-      message: 'hello',
-    },
-  };
-}
-
-export default function DeploymentsLayout({ message, children }) {
-  return (
-    <>
-      <h2>Deployments {message}</h2>
-      {children}
-    </>
-  );
-}
