+++ packages/next/test/fixtures/00-app-dir/app/dashboard/index/lazy.js
@@ -1,3 +1,5 @@
+'client';
+
 export default function LazyComponent() {
   return (
     <>
