+++ packages/next/test/fixtures/00-app-dir/app/dynamic/[category]/[id]/layout.js
@@ -1,11 +1,3 @@
-export async function getServerSideProps({ params }) {
-  return {
-    props: {
-      params,
-    },
-  };
-}
-
 export default function IdLayout({ children, params }) {
   return (
     <>
