+++ packages/next/test/fixtures/00-app-dir/app/dynamic/[category]/[id]/page.js
@@ -1,11 +1,3 @@
-export async function getServerSideProps({ params }) {
-  return {
-    props: {
-      params,
-    },
-  };
-}
-
 export default function IdPage({ children, params }) {
   return (
     <>
