+++ packages/next/test/fixtures/00-app-dir/app/layout.js
@@ -1,0 +1,14 @@
+export default function Root({ children }) {
+  return (
+    <html className="this-is-the-document-html">
+      <head>
+        <title>{`hello world`}</title>
+      </head>
+      <body className="this-is-the-document-body">{children}</body>
+    </html>
+  );
+}
+
+export const config = {
+  revalidate: 0,
+};
