+++ packages/next/test/fixtures/00-app-dir/app/partial-match-[id]/page.js
@@ -1,0 +1,7 @@
+export default function DeploymentsPage(props) {
+  return (
+    <>
+      <p>hello from app/partial-match-[id]. ID is: {props.params.id}</p>
+    </>
+  );
+}
