--- packages/next/test/fixtures/00-app-dir/app/partial-match-[id]/page.server.js
@@ -1,15 +1,0 @@
-export async function getServerSideProps({ params }) {
-  return {
-    props: {
-      id: params.id,
-    },
-  };
-}
-
-export default function DeploymentsPage(props) {
-  return (
-    <>
-      <p>hello from app/partial-match-[id]. ID is: {props.id}</p>
-    </>
-  );
-}
