+++ packages/next/test/fixtures/00-app-dir/app/should-not-serve-client/page.js
@@ -1,3 +1,5 @@
+'client';
+
 export default function ShouldNotServeClientDotJs(props) {
   return (
     <>
