+++ packages/next/test/fixtures/00-app-dir/app/should-not-serve-server/page.js
