+++ packages/next/test/fixtures/00-app-dir/app/slow-layout-and-page-with-loading/slow/page.js
@@ -1,0 +1,13 @@
+import { experimental_use as use } from 'react';
+
+async function getData() {
+  await new Promise(resolve => setTimeout(resolve, 5000));
+  return {
+    message: 'hello from slow page',
+  };
+}
+
+export default function SlowPage(props) {
+  const data = use(getData());
+  return <h1 id="slow-page-message">{data.message}</h1>;
+}
