--- packages/next/test/fixtures/00-app-dir/app/slow-layout-and-page-with-loading/slow/page.server.js
@@ -1,12 +1,0 @@
-export async function getServerSideProps() {
-  await new Promise(resolve => setTimeout(resolve, 5000));
-  return {
-    props: {
-      message: 'hello from slow page',
-    },
-  };
-}
-
-export default function SlowPage(props) {
-  return <h1 id="slow-page-message">{props.message}</h1>;
-}
