+++ packages/next/test/fixtures/00-app-dir/app/slow-layout-with-loading/slow/layout.js
@@ -1,0 +1,18 @@
+import { experimental_use as use } from 'react';
+
+async function getData() {
+  await new Promise(resolve => setTimeout(resolve, 5000));
+  return {
+    message: 'hello from slow layout',
+  };
+}
+
+export default function SlowLayout(props) {
+  const data = use(getData());
+  return (
+    <>
+      <p id="slow-layout-message">{data.message}</p>
+      {props.children}
+    </>
+  );
+}
