+++ packages/next/test/fixtures/00-app-dir/app/slow-layout-with-loading/slow/page.js
